LayoutTests/http/tests/security/block-iframe-to-all-zero-address.html

<html>
<body>
<iframe src="http://0.0.0.0:8000/security/resources/target.html"></iframe></br>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
    }
</script>
<p>This attempts to load an iframe from the all-zeroes address, which should be blocked.</p>
</body>
</html>
